LPS-91609 Refactor move the method to the client focused manager class
