+++ modules/apps/change-tracking/change-tracking-api/src/main/java/com/liferay/change/tracking/CTEngineManager.java
@@ -77,15 +77,6 @@ public interface CTEngineManager {
 	public void enableChangeTracking(long companyId, long userId);
 
 	/**
-	 * Returns the active change tracking collection associated with the given
-	 * user in the scope of the given company.
-	 *
-	 * @param  userId the primary key of the user
-	 * @return the selected change tracking collection
-	 */
-	public Optional<CTCollection> getActiveCTCollectionOptional(long userId);
-
-	/**
 	 * Returns all the change entries associated with the given change
 	 * collection what collide with any of the production change collection
 	 * entries.
@@ -126,18 +117,6 @@ public interface CTEngineManager {
 	public List<CTCollection> getCTCollections(long companyId);
 
 	/**
-	 * Returns all the non production change tracking collection associated
-	 * with the given company.
-	 *
-	 * @param  companyId the primary key of the company
-	 * @param  queryDefinition the object contains settings regarding
-	 *         pagination, order and filter
-	 * @return the list of change tracking collections
-	 */
-	public List<CTCollection> getNonProductionCTCollections(
-		long companyId, QueryDefinition<CTCollection> queryDefinition);
-
-	/**
 	 * Returns all the change entries associated with the given change
 	 * collection.
 	 *
@@ -168,6 +147,18 @@ public interface CTEngineManager {
 	public int getCTEntriesCount(long ctCollectionId);
 
 	/**
+	 * Returns all the non production change tracking collection associated
+	 * with the given company.
+	 *
+	 * @param  companyId the primary key of the company
+	 * @param  queryDefinition the object contains settings regarding
+	 *         pagination, order and filter
+	 * @return the list of change tracking collections
+	 */
+	public List<CTCollection> getNonproductionCTCollections(
+		long companyId, QueryDefinition<CTCollection> queryDefinition);
+
+	/**
 	 * Returns the special change tracking collection which is called production
 	 * and contains all the changes published before.
 	 *
@@ -176,6 +167,13 @@ public interface CTEngineManager {
 	 */
 	public Optional<CTCollection> getProductionCTCollectionOptional(
 		long companyId);
+
+	/**
+	 * Returns the recent change tracking collection id for a specific user.
+	 * @param userId the user id of the user
+	 * @return the recent change tracking collection id
+	 */
+	public long getRecentCTCollectionId(long userId);
 
 	/**
 	 * Returns <code>true</code> if the change tracking is enabled in the scope
