+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/CTManagerImpl.java
@@ -43,7 +43,6 @@ import com.liferay.portal.kernel.transaction.TransactionInvokerUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
@@ -74,7 +73,7 @@ public class CTManagerImpl implements CTManager {
 		}
 
 		Optional<CTCollection> activeCTCollectionOptional =
-			_ctEngineManager.getActiveCTCollectionOptional(userId);
+			getActiveCTCollectionOptional(userId);
 
 		if (!activeCTCollectionOptional.isPresent()) {
 			return Optional.empty();
@@ -142,14 +141,8 @@ public class CTManagerImpl implements CTManager {
 	public Optional<CTEntry> getActiveCTCollectionCTEntryOptional(
 		long userId, long classNameId, long classPK) {
 
-		long companyId = _getCompanyId(userId);
-
-		if (companyId <= 0) {
-			return Optional.empty();
-		}
-
 		Optional<CTCollection> ctCollectionOptional =
-			_ctEngineManager.getActiveCTCollectionOptional(userId);
+			getActiveCTCollectionOptional(userId);
 
 		long ctCollectionId = ctCollectionOptional.map(
 			CTCollection::getCtCollectionId
@@ -157,10 +150,43 @@ public class CTManagerImpl implements CTManager {
 			0L
 		);
 
+		long companyId = _getCompanyId(userId);
+
 		CTEntry ctEntry = _getCTentry(
 			companyId, ctCollectionId, classNameId, classPK);
 
 		return Optional.ofNullable(ctEntry);
+	}
+
+	@Override
+	public Optional<CTCollection> getActiveCTCollectionOptional(long userId) {
+		long companyId = _getCompanyId(userId);
+
+		if (companyId <= 0) {
+			return Optional.empty();
+		}
+
+		if (!_ctEngineManager.isChangeTrackingEnabled(companyId)) {
+			return Optional.empty();
+		}
+
+		long recentCTCollectionId = _ctEngineManager.getRecentCTCollectionId(
+			userId);
+
+		if (recentCTCollectionId == 0L) {
+			Optional<CTCollection> productionCTCollectionOptional =
+				_ctEngineManager.getProductionCTCollectionOptional(companyId);
+
+			recentCTCollectionId = productionCTCollectionOptional.map(
+				CTCollection::getCtCollectionId
+			).orElse(
+				0L
+			);
+
+			_ctEngineManager.checkoutCTCollection(userId, recentCTCollectionId);
+		}
+
+		return _ctEngineManager.getCTCollectionOptional(recentCTCollectionId);
 	}
 
 	@Override
@@ -232,18 +258,8 @@ public class CTManagerImpl implements CTManager {
 		long userId, long resourcePrimKey,
 		QueryDefinition<CTEntry> queryDefinition) {
 
-		long companyId = _getCompanyId(userId);
-
-		if (companyId <= 0) {
-			return Collections.emptyList();
-		}
-
-		if (!_ctEngineManager.isChangeTrackingEnabled(companyId)) {
-			return Collections.emptyList();
-		}
-
 		Optional<CTCollection> ctCollectionOptional =
-			_ctEngineManager.getActiveCTCollectionOptional(userId);
+			getActiveCTCollectionOptional(userId);
 
 		long ctCollectionId = ctCollectionOptional.map(
 			CTCollection::getCtCollectionId
@@ -267,7 +283,7 @@ public class CTManagerImpl implements CTManager {
 		}
 
 		Optional<CTCollection> ctCollectionOptional =
-			_ctEngineManager.getActiveCTCollectionOptional(userId);
+			getActiveCTCollectionOptional(userId);
 
 		long ctCollectionId = ctCollectionOptional.map(
 			CTCollection::getCtCollectionId
@@ -384,10 +400,6 @@ public class CTManagerImpl implements CTManager {
 
 		long companyId = _getCompanyId(userId);
 
-		if (companyId <= 0) {
-			return Optional.empty();
-		}
-
 		if (!_ctEngineManager.isChangeTrackingEnabled(companyId) ||
 			!_ctEngineManager.isChangeTrackingSupported(
 				companyId, classNameId)) {
@@ -396,7 +408,7 @@ public class CTManagerImpl implements CTManager {
 		}
 
 		Optional<CTCollection> ctCollectionOptional =
-			_ctEngineManager.getActiveCTCollectionOptional(userId);
+			getActiveCTCollectionOptional(userId);
 
 		if (!ctCollectionOptional.isPresent()) {
 			return Optional.empty();
